requirements.txt
click
openai-whisper
requests
